+++ chainer_chemistry/saliency/calculator/__init__.py
@@ -1,5 +1,5 @@
 from chainer_chemistry.saliency.calculator import base_calculator  # NOQA
-from chainer_chemistry.saliency.calculator import common  # NOQA
+from chainer_chemistry.saliency.calculator import calculator_utils  # NOQA
 from chainer_chemistry.saliency.calculator import gradient_calculator  # NOQA
 from chainer_chemistry.saliency.calculator import integrated_gradients_calculator  # NOQA
 from chainer_chemistry.saliency.calculator import occlusion_calculator  # NOQA
@@ -9,4 +9,4 @@ from chainer_chemistry.saliency.calculator.gradient_calculator import GradientCa
 from chainer_chemistry.saliency.calculator.integrated_gradients_calculator import IntegratedGradientsCalculator  # NOQA
 from chainer_chemistry.saliency.calculator.occlusion_calculator import OcclusionCalculator  # NOQA
 
-from chainer_chemistry.saliency.calculator.common import GaussianNoiseSampler  # NOQA
+from chainer_chemistry.saliency.calculator.calculator_utils import GaussianNoiseSampler  # NOQA
