+++ chainer_chemistry/saliency/visualizer/__init__.py
@@ -1,5 +1,5 @@
 from chainer_chemistry.saliency.visualizer import base_visualizer  # NOQA
-from chainer_chemistry.saliency.visualizer import common  # NOQA
+from chainer_chemistry.saliency.visualizer import visualizer_utils  # NOQA
 from chainer_chemistry.saliency.visualizer import image_visualizer  # NOQA
 from chainer_chemistry.saliency.visualizer import mol_visualizer  # NOQA
 from chainer_chemistry.saliency.visualizer import table_visualizer  # NOQA
@@ -10,7 +10,7 @@ from chainer_chemistry.saliency.visualizer.mol_visualizer import MolVisualier  #
 from chainer_chemistry.saliency.visualizer.mol_visualizer import SmilesVisualizer  # NOQA
 from chainer_chemistry.saliency.visualizer.table_visualizer import TableVisualizer  # NOQA
 
-from chainer_chemistry.saliency.visualizer.common import abs_max_scaler  # NOQA
-from chainer_chemistry.saliency.visualizer.common import min_max_scaler  # NOQA
-from chainer_chemistry.saliency.visualizer.common import normalize_scaler  # NOQA
-from chainer_chemistry.saliency.visualizer.common import red_blue_cmap  # NOQA
+from chainer_chemistry.saliency.visualizer.visualizer_utils import abs_max_scaler  # NOQA
+from chainer_chemistry.saliency.visualizer.visualizer_utils import min_max_scaler  # NOQA
+from chainer_chemistry.saliency.visualizer.visualizer_utils import normalize_scaler  # NOQA
+from chainer_chemistry.saliency.visualizer.visualizer_utils import red_blue_cmap  # NOQA
