+++ chainer_chemistry/saliency/visualizer/image_visualizer.py
@@ -7,7 +7,7 @@ import matplotlib.cm as cm
 from chainer import cuda
 
 from chainer_chemistry.saliency.visualizer.base_visualizer import BaseVisualizer  # NOQA
-from chainer_chemistry.saliency.visualizer.common import abs_max_scaler  # NOQA
+from chainer_chemistry.saliency.visualizer.visualizer_utils import abs_max_scaler  # NOQA
 
 
 class ImageVisualizer(BaseVisualizer):
